refactor: move is*Connection() to connection.ts
Moves the functions to src/auth/connections.ts
an updates all imports
