+++ src/codewhisperer/util/showSsoPrompt.ts
@@ -13,9 +13,10 @@ import { isValidResponse } from '../../shared/wizards/wizard'
 import { CancellationError } from '../../shared/utilities/timeoutUtils'
 import { ToolkitError } from '../../shared/errors'
 import { createCommonButtons } from '../../shared/ui/buttons'
-import { createBuilderIdItem, createIamItem, createSsoItem, isIamConnection } from '../../auth/auth'
+import { createBuilderIdItem, createIamItem, createSsoItem } from '../../auth/auth'
 import { telemetry } from '../../shared/telemetry/telemetry'
 import { isCloud9 } from '../../shared/extensionUtilities'
+import { isIamConnection } from '../../auth/connection'
 
 export const showConnectionPrompt = async () => {
     // Skip this prompt on C9 because:
