+++ src/test/codewhisperer/util/authUtil.test.ts
@@ -4,13 +4,12 @@
  */
 
 import * as assert from 'assert'
-import { isSsoConnection } from '../../../auth/auth'
 import { AuthUtil } from '../../../codewhisperer/util/authUtil'
 import { getTestWindow } from '../../shared/vscode/window'
 import { SeverityLevel } from '../../shared/vscode/message'
 import { createBuilderIdProfile, createSsoProfile, createTestAuth } from '../../credentials/testUtil'
 import { captureEventOnce } from '../../testUtil'
-import { codewhispererScopes } from '../../../auth/connection'
+import { codewhispererScopes, isSsoConnection } from '../../../auth/connection'
 
 const enterpriseSsoStartUrl = 'https://enterprise.awsapps.com/start'
